IMPROVED EDGE GENERATION
Edges are now generated by looking at an edge's degree and comparing it
to it's current number of edges. This is done by generating a random
degree for each node, and suming the node's entry in the adjacency
matrix (adjMatrix). Once the adjacencey is established the edge is
created and the sum recalculated.

This is possible because of the nature of an adjacencey matrix. Since
the matrix is made up of 1's or 0's, the sum of each row is that node's
degree.

Generating the edges this way no loger reqires that the adjMatrix be
initalised to -1's, and it can instead be initalised to 0's.

This commit also inclued a function for returning an array of the
summation of each of the rows of a matrix. This is uesed to determin the
current degree of a node.