GUI START
The program now has an interactive GUI. The GUI is uesd to set the
desired number of nodes and create the graph. The program does now
require the controlP5 library for processing. You can find instructions
on how to install the library on the processing wiki, here:
github.com/processing/processing/wiki/How-to-Install-a-Contributed-Library