Added function to find each nodes exctricity
The excentricity of a node is the minimum distance between it and the
node furthest away from it. I find it by finding the distance k matrix
where k is the size of the graph. Once that matrix is found the largest
value in each row or colume is the excentricity of the respective node.

If there are any 0 in this graph it is a disconnected graph.